app/requirements.txt
flask
mysql-connector-python
flask-paginate
pandas
tmdbsimple
rottentomatoes-python
rotten_tomatoes_scraper
